+++ test/utils/feature_manager_test.dart
@@ -4,10 +4,14 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 void main() {
   late FeatureManager _target;
+  late SharedPreferences sharedPreferences;
 
   setUp(() async {
     SharedPreferences.setMockInitialValues(<String, Object>{});
-    _target = FeatureManager();
+    sharedPreferences = await SharedPreferences.getInstance();
+    _target = FeatureManager(
+      sharedPreferences: sharedPreferences,
+    );
   });
 
   group('isEnabled', () {
@@ -16,6 +20,10 @@ void main() {
       SharedPreferences.setMockInitialValues(<String, Object>{
         'key': true,
       });
+      sharedPreferences = await SharedPreferences.getInstance();
+      _target = FeatureManager(
+        sharedPreferences: sharedPreferences,
+      );
       final Feature feature = Feature(
         key: 'key',
         title: '',
@@ -24,14 +32,14 @@ void main() {
       );
 
       // when
-      final bool value = await _target.isEnabled(feature);
+      final bool value = _target.isEnabled(feature);
 
       // then
       expect(value, true);
     });
 
     test('when feature is toggle and value is null should return default value',
-        () async {
+        () {
       // given
       final Feature feature = Feature(
         key: 'key',
@@ -42,7 +50,7 @@ void main() {
       );
 
       // when
-      final bool value = await _target.isEnabled(feature);
+      final bool value = _target.isEnabled(feature);
 
       // then
       expect(value, true);
@@ -50,7 +58,7 @@ void main() {
 
     test(
         'when feature is toggle and value is null and default value is null should return false',
-        () async {
+        () {
       // given
       final Feature feature = Feature(
         key: 'key',
@@ -60,13 +68,13 @@ void main() {
       );
 
       // when
-      final bool value = await _target.isEnabled(feature);
+      final bool value = _target.isEnabled(feature);
 
       // then
       expect(value, false);
     });
 
-    test('when feature is not toggle should return false', () async {
+    test('when feature is not toggle should return false', () {
       // given
       final Feature feature = Feature(
         key: 'key',
@@ -76,7 +84,7 @@ void main() {
       );
 
       // when
-      final bool value = await _target.isEnabled(feature);
+      final bool value = _target.isEnabled(feature);
 
       // then
       expect(value, false);
@@ -89,6 +97,10 @@ void main() {
       SharedPreferences.setMockInitialValues(<String, Object>{
         'key': 'some text',
       });
+      sharedPreferences = await SharedPreferences.getInstance();
+      _target = FeatureManager(
+        sharedPreferences: sharedPreferences,
+      );
       final Feature feature = Feature(
         key: 'key',
         title: '',
@@ -97,13 +109,13 @@ void main() {
       );
 
       // when
-      final String value = await _target.getValue(feature) as String;
+      final String value = _target.getValue(feature) as String;
 
       // then
       expect(value, 'some text');
     });
 
-    test('when stored value is null should return default value', () async {
+    test('when stored value is null should return default value', () {
       // given
       final Feature feature = Feature(
         key: 'key',
@@ -114,7 +126,7 @@ void main() {
       );
 
       // when
-      final String value = await _target.getValue(feature) as String;
+      final String value = _target.getValue(feature) as String;
 
       // then
       expect(value, 'default text');