+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+* BREAKING: now you need to provide SharedPreferences instance to create feature manager
+* Added sync calls for FeatureManager
+* Added additional getters for FeatureManger
+* Added json type for Feature
+
 ## 1.0.0
 
 * Initial release of Feature Manager